backend/requirements.txt
Django==5.2
debugpy==1.8.20
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
django-filter==25.2
uvicorn[standard]==0.41.0
gunicorn==23.0.0

# Testing & linting
pytest==8.4.1
pytest-django==4.11.1
pytest-cov==6.2.1
factory-boy==3.3.3
faker==37.3.0
flake8==7.3.0
black==25.1.0
isort==6.0.1

# Error tracking (optional — set SENTRY_DSN env var to enable)
sentry-sdk[django]==2.22.0
